update based on comments and add ut

+++ python/chronos/src/bigdl/chronos/model/autoformer/Autoformer.py
@@ -157,8 +157,7 @@ class AutoFormer(pl.LightningModule):
 
     def validation_step(self, batch, batch_idx):
         batch_x, batch_y, batch_x_mark, batch_y_mark = map(lambda x: x.float(), batch)
-        outputs = self(batch_x.float(), batch_x_mark.float(), batch_y,
-                       batch_y_mark.float())
+        outputs = self(batch_x, batch_x_mark, batch_y, batch_y_mark)
 
         outputs = outputs[:, -self.pred_len:, :]
         batch_y = batch_y[:, -self.pred_len:, :]
@@ -166,8 +165,8 @@ class AutoFormer(pl.LightningModule):
 
     def predict_step(self, batch, batch_idx):
         batch_x, batch_y, batch_x_mark, batch_y_mark = map(lambda x: x.float(), batch)
-        outputs = self(batch_x.float(), batch_x_mark.float(), batch_y,
-                       batch_y_mark.float())
+        outputs = self(batch_x, batch_x_mark, batch_y, batch_y_mark)
+        
         outputs = outputs[:, -self.pred_len:, :]
         return outputs
